requirements.txt
pywebview
blinkpy
Flask >= 0.12
